refactor: RunningCompletedEvent 패키지 이동
