+++ tamago-core/src/main/kotlin/tamago/server/core/running/domain/event/RunningCompletedEvent.kt
@@ -1,4 +1,4 @@
-package tamago.server.core.common.event
+package tamago.server.core.running.domain.event
 
 import tamago.server.core.common.vo.UserId
 import java.time.LocalDateTime
